% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnfparser.R
\name{is.bnf_grammar}
\alias{is.bnf_grammar}
\title{Validate BNF Grammar}
\usage{
is.bnf_grammar(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a valid BNF grammar object, FALSE otherwise
}
\description{
Checks if an object is a valid BNF grammar
}
