% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnfparser.R
\name{print.bnf_grammar}
\alias{print.bnf_grammar}
\title{Print Method for BNF Grammar Objects}
\usage{
\method{print}{bnf_grammar}(x, ...)
}
\arguments{
\item{x}{An object of class "bnf_grammar"}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for BNF Grammar Objects
}
