% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnfparser.R
\name{summary.bnf_grammar}
\alias{summary.bnf_grammar}
\title{Summary Method for BNF Grammar Objects}
\usage{
\method{summary}{bnf_grammar}(object, ...)
}
\arguments{
\item{object}{An object of class "bnf_grammar"}

\item{...}{Additional arguments passed to summary}
}
\value{
Invisibly returns the input object
}
\description{
Summary Method for BNF Grammar Objects
}
